[runtime] Move simulated network infrastructure from p2p to runtime
Migrate the link, bandwidth, and transmitter modules from p2p::simulated
to runtime::simulated, making them generic over peer and channel types.
This allows the network simulation infrastructure to be reused across
different contexts (e.g., the multihead runtime with Ipv4Addr peers).

Changes:
- Add runtime/src/simulated/ module with Link, bandwidth, and transmitter
- Make transmitter generic over peer type P and channel type C
- Re-export runtime's Link from p2p::simulated::ingress
- Update p2p::simulated::transmitter to use type aliases wrapping runtime
- Delete p2p/src/simulated/bandwidth.rs (now in runtime)
- Simplify multihead module by removing incompatible LinkAwareNetwork wrapper
- Add required dependencies (num-rational, num-traits, rand_distr) to runtime
